pyproject.toml
[project]
name = "aristotle-system"
version = "0.1.0"
description = "The core system for aristotle to provide accurate information regarding codebases, code repositories, and packages"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.11.0",
    "datasets>=4.3.0",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.119.1",
    "graphiti-core[sentence-transformers]>=0.21.0",
    "langchain>=1.0.2",
    "langchain-community>=0.4",
    "langchain-core>=1.0.0",
    "langchain-ollama>=1.0.0",
    "matplotlib>=3.10.6",
    "nbconvert>=7.16.6",
    "networkx>=3.5",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "pydantic>=2.12.3",
    "pygit2>=1.18.2",
    "pytest>=8.4.2",
    "python-frontmatter>=1.1.0",
    "ragas>=0.3.7",
    "requests>=2.32.5",
    "tqdm>=4.67.1",
    "uuid>=1.30",
    "uvicorn>=0.38.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/aristotle"]
